move AotTargetDevice to aot_module_builder.h
