+++ taichi/program/aot_module_builder.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 #include "taichi/ir/snode.h"
-
+#include "taichi/backends/device.h"
 namespace taichi {
 namespace lang {
 
@@ -61,6 +61,51 @@ class AotModuleBuilder {
   static bool all_fields_are_dense_in_container(const SNode *container);
 
   static int find_children_id(const SNode *snode);
+};
+
+// Only responsible for reporting device capabilities
+class AotTargetDevice : public Device {
+ public:
+  AotTargetDevice(Arch arch) {
+    // TODO: make this configurable
+    set_default_caps(arch);
+  }
+
+  void set_default_caps(Arch arch) {
+    if (arch == Arch::vulkan) {
+      set_cap(DeviceCapability::spirv_version, 0x10300);
+    }
+  }
+
+  DeviceAllocation allocate_memory(const AllocParams &params) override {
+    TI_NOT_IMPLEMENTED;
+  }
+  void dealloc_memory(DeviceAllocation handle) override {
+    TI_NOT_IMPLEMENTED;
+  }
+  std::unique_ptr<Pipeline> create_pipeline(
+      const PipelineSourceDesc &src,
+      std::string name = "Pipeline") override {
+    TI_NOT_IMPLEMENTED;
+  }
+  void *map_range(DevicePtr ptr, uint64_t size) override {
+    TI_NOT_IMPLEMENTED;
+  }
+  void *map(DeviceAllocation alloc) override {
+    TI_NOT_IMPLEMENTED;
+  }
+  void unmap(DevicePtr ptr) override {
+    TI_NOT_IMPLEMENTED;
+  }
+  void unmap(DeviceAllocation alloc) override {
+    TI_NOT_IMPLEMENTED;
+  }
+  void memcpy_internal(DevicePtr dst, DevicePtr src, uint64_t size) override {
+    TI_NOT_IMPLEMENTED;
+  }
+  Stream *get_compute_stream() override {
+    TI_NOT_IMPLEMENTED;
+  }
 };
 
 }  // namespace lang
